% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_rate.R
\name{compute_hr}
\alias{compute_hr}
\title{Compute Hit Rate at k}
\usage{
compute_hr(test_pred, k, user_col = user, label_col = label, pred_col = pred)
}
\arguments{
\item{test_pred}{A data frame containing predictions and truth labels for a test data set.}

\item{k}{An integer specifying HR at k.}

\item{user_col}{Name of the column containing the user id's. Default: user}

\item{label_col}{Name of the column containing binary truth labels indicating positive/like (1) or
negative/dislike (0) for each user-item combination. Default: label}

\item{pred_col}{Name of the column containing model prediction scores for each user-item combination. Default: pred}
}
\value{
HR@k:  a decimal number in the range [0,1].
}
\description{
This function is available only for backward compatibility. Superseded by hit_rate().
}
\examples{
test_pred <- data.frame(pred = runif(5), user = 1:5, item = 1:5, label = sample(0:1, 5, replace=TRUE))
compute_hr(test_pred, k = 2)
test_pred <- data.frame(prediction = runif(5), user_id = 1:5, item_id = 1:5, label = sample(0:1, 5, replace=TRUE))
compute_hr(test_pred, k = 2, user_col=user_id, pred_col=prediction)
}
