% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity}
\alias{diversity}
\title{Compute diversity}
\usage{
diversity(
  x,
  k,
  method = c("jaccard", "spearman"),
  agg_fun = mean,
  no_aggregate = FALSE,
  pred_col = "pred",
  user_col = "user",
  item_col = "item",
  ...
)
}
\arguments{
\item{x}{data.frame that comprises a column of predictions (identified
by \code{pred_col}), a column of users (identified by \code{user_col})
and a column of items (identified by \code{item_col})}

\item{k}{numeric scalar that specifies the length of recommendation list}

\item{method}{character scalar that specifies the measure of dissimilarity
to employ}

\item{agg_fun}{function that aggregates the pairwise dissimilarities}

\item{no_aggregate}{logical scalar that specifies to return the
un-aggregated pairwise dissimilarities instead of a single aggregate value}

\item{pred_col}{string that identifies \code{x}'s column of predictions}

\item{user_col}{string that identifies \code{x}'s column of users}

\item{item_col}{string that identifies \code{x}'s column of items}

\item{...}{arguments to pass to dissimilarity function}
}
\value{
numeric scalar if \code{no_aggregate} is FALSE; otherwise numeric
vector of length \eqn{n}-choose-two, where \eqn{n} is the number of unique
values in \code{x[[user_col]]}.
}
\description{
Compute the diversity of a collection of recommendation lists.  `Diversity'
here is defined as aggregate pairwise dissimilarity, where the aggregation
function and the measure of dissimilarity are specified by the user via
the parameters named `agg_fun ' and `method'.
}
\examples{
# Generate data.
set.seed(1L)
nuser <- 10L                        # number of unique users
nitem <- 50L                        # number of unique items
u <- paste0("u", seq_len(nuser))    # set of user IDs
i <- paste0("i", seq_len(nitem))    # set of item IDs
x <- expand.grid(pred = NA_real_, item = i, user = u)
x$pred <- unlist(replicate(nuser, { # predictions vector
    y <- runif(nitem)               # random values in [0,1]...
    y / sum(y)                      # ...that sum to one
}, simplify = FALSE))
k <- 5L                          # length of each recommendation list
# Compute diversity of data multiple ways for comparison.
measures <- c("jaccard", "spearman")
names(measures) <- measures
d <- lapply(measures, function (m)
    diversity(x, k = k, method = m, no_aggregate = TRUE))
# Plot diversity values.
boxplot(d, xlab = "measure", ylab = "diversity",
    main = "Diversity by measure of dissimilarity")
}
\seealso{
\code{evalrs:::spearman_dist}
}
