% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_sampled}
\alias{diversity_sampled}
\title{Estimate diversity from random samples of users}
\usage{
diversity_sampled(x, nrep = 3L, maxuser = 100L, user_col = "user", ...)
}
\arguments{
\item{x}{data.frame that comprises a column of predictions (identified
by \code{pred_col}), a column of users (identified by \code{user_col})
and a column of items (identified by \code{item_col})}

\item{nrep}{number of times to compute diversity of recommendation system
for the sample of users of the specified size}

\item{maxuser}{maximum number of users to include in random samples;
samples of size \code{min(maxuser, n)} are used, where \code{n} is the
number of unique users found in \code{x}}

\item{user_col}{scalar that identifies \code{x}'s column of users}

\item{...}{arguments passed to \code{evalrs::diversity}, some of which (such
as \code{nrec}) are required}
}
\value{
mean of the diversity values calculated from subsets of the data
}
\description{
Calculating the diversity of a recommendation system by way of
\code{evarls::diversity} can be onerous, as it entails \eqn{n\choose2}
comparisons of recommendation lists, where \eqn{n} is the number of users
for which the system has recommendations.  The burden is especially
great for diversity calculations based on Kendall distance, which for
two lists of length \eqn{k} nominally entails at least \eqn{k\choose2}
comparisons of its own.  In light thereof, the present function serves
to reduce the computational burden by estimating the result from random
samples of users of a specified size.
}
\examples{
# Generate data.
set.seed(1L)
nuser <- 1000L                      # number of unique users
nitem <- 5000L                      # number of unique items
u <- paste0("u", seq_len(nuser))    # set of user IDs
i <- paste0("i", seq_len(nitem))    # set of item IDs
x <- expand.grid(pred = NA_real_, item = i, user = u)
x$pred <- unlist(replicate(nuser, { # predictions vector
    y <- runif(nitem)               # random values in [0,1]...
    y / sum(y)                      # ...that sum to one
}, simplify = FALSE))
nrec <- 5L                          # length of each recommendation list
# Estimate diversity from the default number of random samples of users of
# the default size.
diversity_sampled(x, nrec = nrec, method = "kendall")
}
\seealso{
\code{evalrs::diversity}
}
