% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalrs.R
\docType{package}
\name{evalrs}
\alias{evalrs}
\title{evalrs: Evaluation methods for recommender systems}
\description{
The evalrs package provides functionality for computing evaluation metrics
for recommender systems.
}
\section{evalrs functions}{

hit_rate,
mean_avg_precision,
mean_recip_rank,
mpr_score,
ndcg,
novelty
}

\section{Author}{

Amanda Peterson
}

