% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_avg_precision.R
\name{mean_avg_precision}
\alias{mean_avg_precision}
\title{Mean Average Precision (MAP@k)}
\usage{
mean_avg_precision(
  test_pred,
  k,
  user_col = user,
  label_col = label,
  pred_col = pred
)
}
\arguments{
\item{test_pred}{A data frame containing predictions for the test data set.}

\item{k}{An integer specifying MAP at k.}

\item{user_col}{Name of the test_pred column containing the user id's. Default: user}

\item{label_col}{Name of the test_pred column containing binary truth labels indicating positive/like (1) or
negative/dislike (0) for each user-item combination. Default: label}

\item{pred_col}{Name of the test_pred column containing model prediction scores for each user-item combination. Default: pred}
}
\value{
MAP@k, which is a decimal number in the range [0,1].
}
\description{
Compute the mean average precision for a top-k list of recommendations.
}
\details{
Average precision at k (for each user) is defined as \deqn{AP@k = \frac{1}{m} \sum_{i=1}^k P(i) * rel(i)}
where \eqn{P(i)} is the precision at i. \eqn{rel(i) = 1} if the item is relevant (or liked by user)
and 0 if not relevant (not liked by user). \eqn{m = min(k, number of positives in full item space)}.

Mean average precision at k is the mean of AP@k over all users.
}
\examples{
test_pred <- data.frame(pred = runif(20), user = rep(1:5,each=4), item = as.vector(sapply(rep(1,5), FUN=function(x){ rep(sample(1:4), x) } )), label = as.vector(sapply(rep(4,5), FUN=function(x){ sample(c(1,rep(0,3)), x, replace=FALSE) })))
mean_avg_precision(test_pred, k = 2)
test_pred <- data.frame(prediction = runif(20), user_id = rep(1:5,each=4), item_id = as.vector(sapply(rep(1,5), FUN=function(x){ rep(sample(1:4), x) } )), label = as.vector(sapply(rep(4,5), FUN=function(x){ sample(c(1,rep(0,3)), x, replace=FALSE) })))
mean_avg_precision(test_pred, k = 2, user_col=user_id, pred_col=prediction)
}
\seealso{
http://sdsawtelle.github.io/blog/output/mean-average-precision-MAP-for-recommender-systems.html
}
