% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpr_score.R
\name{pctl_rank}
\alias{pctl_rank}
\title{Compute percentile-ranks of the given sample.}
\usage{
pctl_rank(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector of length \code{length(x)} of numbers in the interval
\eqn{[0,100]}
}
\description{
Compute percentile-ranks of the given sample.
}
