% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman.dist.R
\name{spearman_dist}
\alias{spearman_dist}
\title{Compute Spearman's footrule distance with location parameter \eqn{L}}
\usage{
spearman_dist(x, y, L = 1L + length(x), normalize = TRUE)
}
\arguments{
\item{x}{named numeric vector of rankings of length \code{length(y)}}

\item{y}{named numeric vector of rankings of length \code{length(x)}}

\item{L}{numeric scalar to impute to missing elements}

\item{normalize}{logical scalar that indicates whether to map the result to
the interval [0,1]}
}
\value{
numeric scalar
}
\description{
Compute the distance between two rankings based Spearman's footrule metric
as proposed by Fagin et al. (2003) for top-k lists.  Spearman's footrule
metric with location parameter \eqn{L} yields the L1 distance between two
permutations after having modified them such that elements present in
exactly one of the two are assigned the value \eqn{L} in the other.
}
\details{
If argument \code{normalize} equals \code{TRUE} then the Spearman's footrule
distance with location parameter \code{L} between arguments \code{x} and
\code{y} each of length \eqn{k} is mapped to the interval \eqn{[0,1]} by
dividing by \code{k * (2 * max(L, k + 1) - (k + 1))}.
}
\examples{
# Randomly rank k random letters.  Pretend that these k elements are the k
# of highest rank in some larger ranking.  Do it again.  Compare the two
# top-k lists via spearman_dist.
set.seed(1L)
k <- 10L
u <- LETTERS
x <- stats::setNames(sample(k), sample(u, k))
y <- stats::setNames(sample(k), sample(u, k))
spearman_dist(x, y)
}
\references{
Ronald Fagin, Ravi Kumar and D. Sivakumar. (2003). Comparing top k lists.
SIAM J. Discrete Mathematics 17, 1, 134–160.
}
